feat: deep NPC system, shop makeover, and Phase 5 economy features
A large update covering the full NPC epic, a top-to-bottom player-shop
rebuild, and the remaining Phase 5 economy content, plus repo cleanup.

Deep NPC system (all pillars):
- NotchNpcEntity (GeckoLib GeoEntity) with one blank-slate item, ownership-
  gated editing, and a "pick up" that packs full config into the item.
- Tabbed editor: appearance/model picker (vanilla + GeckoLib + entity
  disguise), behaviors (stationary/wander/follow/patrol/guard), pose editor
  with presets + live custom-pose sliders, branching dialogue studio,
  live equipment inventory, stats & abilities screen, and presets.
- Public NotchNpcApi: spawnNpc/spawnNpcFromPreset, preset apply, and a
  custom-role registry; /npc spawn [preset] and /npc debug commands.

Player-shop top-to-bottom makeover:
- New code-drawn browse, manage-hub, and listing-editor screens replacing
  the PNG-textured 6-listing screen; full pagination (all 27 listings),
  coin/barter/combined pricing, immediate per-action saves (no more
  save-on-close), and one-click earnings collection.
- Retired the legacy ShopkeeperEntity + Merchant License + spawn-egg stack;
  the NPC SHOP role is now the only shop path.

Phase 5 economy features:
- Enchanter NPC (repair / buy specific enchants / extract to book, SINK).
- Custom-currency maker (generates a resource pack from admin art + name).
- Waystone teleport fee (soft integration, coin SINK).
- Expanded, price-scaling auction listing fee + sale tax with caps.
- Cosmetics shop (datapack-driven, sells any mod's cosmetics via item or
  unlock command).
- Offline trade offers (escrow + offline-safe delivery via a mailbox).
- Trade history / receipts screen (/receipts), captured via a single
  BalanceStore hook.
- Fixed the /ah cancel escrow bug (bids are now refunded, offline-safe).

Repo cleanup: stop tracking build/.gradle/.idea artifacts and dead source;
ignore local drop-in and borrowed soft-dep API jars.

+++ src/main/java/net/fugginbeenus/notchcurrency/atm/ATMBlock.java
@@ -73,7 +73,7 @@ public class ATMBlock extends HorizontalFacingBlock {
 
         // Server: sync balance to HUD immediately
         if (player instanceof ServerPlayerEntity spe) {
-            int bal = net.fugginbeenus.notchcurrency.core.BalanceStore.get(spe);
+            long bal = net.fugginbeenus.notchcurrency.core.BalanceStore.get(spe);
             net.fugginbeenus.notchcurrency.net.NotchPackets.sendBalance(spe, bal);
         }
